refactor(ava/advisor): split visual-encoder and chart-type-recommend folders
