+++ packages/ava/src/advisor/advise-pipeline/chart-type-recommend/index.ts
@@ -1,0 +1,1 @@
+export * from './get-chart-Type';
