+++ packages/ava/src/advisor/advise-pipeline/spec-generator/charts/area.ts
@@ -1,5 +1,5 @@
 import { hasSubset, intersects } from '../../../utils';
-import { splitAreaXYSeries } from '../splitFields';
+import { splitAreaXYSeries } from '../../visual-encoder/split-fields';
 
 import type { Data } from '../../../../common/types';
 import type { Advice, BasicDataPropertyForAdvice } from '../../../types';
