+++ packages/ava/src/advisor/advise-pipeline/spec-generator/charts/bar.ts
@@ -1,4 +1,4 @@
-import { splitBarXYSeries } from '../splitFields';
+import { splitBarXYSeries } from '../../visual-encoder/split-fields';
 
 import type { Data } from '../../../../common/types';
 import type { Advice, BasicDataPropertyForAdvice } from '../../../types';
