+++ packages/ava/src/advisor/advise-pipeline/spec-generator/charts/pie.ts
@@ -1,4 +1,4 @@
-import { splitAngleColor } from '../splitFields';
+import { splitAngleColor } from '../../visual-encoder/split-fields';
 
 import type { Data } from '../../../../common/types';
 import type { Advice, BasicDataPropertyForAdvice } from '../../../types';
